refactor: move cache keys to session keys for consistency
