--- app/config/constants.js
@@ -1,10 +1,0 @@
-module.exports = {
-  cacheKeys: {
-    answers: {
-      cattle: 'cattle',
-      cattleType: 'cattleType',
-      pigs: 'pigs',
-      sheep: 'sheep'
-    }
-  }
-}
