+++ app/routes/farmer-apply/cattle-type.js
@@ -1,5 +1,5 @@
 const Joi = require('joi')
-const { cacheKeys: { answers } } = require('../../config/constants')
+const { answers } = require('../../session/keys')
 const session = require('../../session')
 
 const backLink = '/farmer-apply/cattle'
